requirements.txt
sqlalchemy>=2.0.0
pydantic>=2.0.0
google-generativeai>=0.3.0
fastmcp>=0.1.0
aioimaplib>=1.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
